# specs/001-crud-empleados/contracts/empleados.openapi.yaml
openapi: 3.0.3
info:
  title: API Empleados
  version: 2.0.0
  description: Contrato CRUD para gestión de empleados.
servers:
  - url: http://localhost:8081
security:
  - basicAuth: []
paths:
  /api/v2/empleados:
    post:
      summary: Crear empleado
      operationId: crearEmpleado
      tags: [Empleados]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequest'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      summary: Listar empleados
      operationId: listarEmpleados
      tags: [Empleados]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Página de empleados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v2/empleados/{clave}:
    parameters:
      - name: clave
        in: path
        required: true
        schema:
          type: string
          pattern: '^EMP-\\d+$'
    get:
      summary: Obtener empleado por clave
      operationId: obtenerEmpleado
      tags: [Empleados]
      responses:
        '200':
          description: Empleado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar empleado
      operationId: actualizarEmpleado
      tags: [Empleados]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoUpdateRequest'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar empleado
      operationId: eliminarEmpleado
      tags: [Empleados]
      responses:
        '204':
          description: Empleado eliminado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    EmpleadoCreateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          maxLength: 100
        direccion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
    EmpleadoUpdateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          maxLength: 100
        direccion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
    EmpleadoResponse:
      type: object
      required: [clave, nombre, direccion, telefono]
      properties:
        clave:
          type: string
          maxLength: 100
        nombre:
          type: string
          maxLength: 100
        direccion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
    EmpleadoPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmpleadoResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

  responses:
    BadRequest:
      description: Datos inválidos
    Unauthorized:
      description: Autenticación requerida o inválida
    NotFound:
      description: Recurso no encontrado